+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @whichly/web
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @whichly/react@2.0.0
+
 ## 0.0.3
 
 ### Patch Changes